fix(sonarcloud): remove duplicate code
