+++ src/features/reports/publisher_records/export_S21/specific_records/field_service_groups/useFieldServiceGroups.tsx
@@ -11,6 +11,7 @@ import { FieldServiceGroupType } from '@definition/field_service_groups';
 import { personsState } from '@states/persons';
 import { FieldServiceGroupsProps } from './index.types';
 import usePersons from '@features/persons/hooks/usePersons';
+import useParentUncheckHandler from '../useParentUncheckHandler';
 
 const useFieldServiceGroups = ({ onExport }: FieldServiceGroupsProps) => {
   const { t } = useAppTranslation();
@@ -122,34 +123,10 @@ const useFieldServiceGroups = ({ onExport }: FieldServiceGroupsProps) => {
     toggledItemRef.current[itemId] = isSelected;
   };
 
-  const findParentIdByItem = (
-    dataSource,
-    itemId: string,
-    parentId: string = null
-  ) => {
-    for (const item of dataSource) {
-      if (item.id === itemId) {
-        return parentId;
-      }
-      if (item.children) {
-        const found = findParentIdByItem(item.children, itemId, item.id);
-        if (found) {
-          return found;
-        }
-      }
-    }
-    return null;
-  };
-
-  const deleteSelectionFromParentItem = (oldSelectedList: string[]) => {
-    const missedItem = selected.filter(
-      (item) => !oldSelectedList.includes(item)
-    )[0];
-
-    const missedItemParent = findParentIdByItem(groups, missedItem);
-
-    return oldSelectedList.filter((item) => item !== missedItemParent);
-  };
+  const { deleteSelectionFromParentItem } = useParentUncheckHandler(
+    groups,
+    selected
+  );
 
   const handleSelectionChange = (newSelectedItems: string[]) => {
     setSelected(newSelectedItems);
