+++ src/features/reports/publisher_records/export_S21/specific_records/inactive_publishers/useInactivePublishers.tsx
@@ -8,6 +8,7 @@ import { fullnameOptionState } from '@states/settings';
 import { buildPersonFullname } from '@utils/common';
 import { InactivePublishersProps } from './index.types';
 import usePersons from '@features/persons/hooks/usePersons';
+import useParentUncheckHandler from '../useParentUncheckHandler';
 
 const useInactivePublishers = ({ onExport }: InactivePublishersProps) => {
   const { t } = useAppTranslation();
@@ -87,34 +88,10 @@ const useInactivePublishers = ({ onExport }: InactivePublishersProps) => {
     toggledItemRef.current[itemId] = isSelected;
   };
 
-  const findParentIdByItem = (
-    dataSource,
-    itemId: string,
-    parentId: string = null
-  ) => {
-    for (const item of dataSource) {
-      if (item.id === itemId) {
-        return parentId;
-      }
-      if (item.children) {
-        const found = findParentIdByItem(item.children, itemId, item.id);
-        if (found) {
-          return found;
-        }
-      }
-    }
-    return null;
-  };
-
-  const deleteSelectionFromParentItem = (oldSelectedList: string[]) => {
-    const missedItem = selected.filter(
-      (item) => !oldSelectedList.includes(item)
-    )[0];
-
-    const missedItemParent = findParentIdByItem(groups, missedItem);
-
-    return oldSelectedList.filter((item) => item !== missedItemParent);
-  };
+  const { deleteSelectionFromParentItem } = useParentUncheckHandler(
+    groups,
+    selected
+  );
 
   const handleSelectionChange = (newSelectedItems: string[]) => {
     setSelected(newSelectedItems);
